Sync Forks

+++ ssc/common.h
@@ -54,8 +54,9 @@
 #include <vector>
 #include "core.h"
 
-#include "lib_weatherfile.h"
-#include "lib_pv_shade_loss_mpp.h"
+#include "../shared/lib_util.h"
+#include "../shared/lib_weatherfile.h"
+#include "../shared/lib_pv_shade_loss_mpp.h"
 
 extern var_info vtab_standard_financial[];
 extern var_info vtab_standard_loan[];
@@ -149,7 +150,7 @@ class weatherdata : public weather_data_provider
 	vec get_vector(var_data *v, const char *name, size_t *len = nullptr);
 	ssc_number_t get_number(var_data *v, const char *name);
 
-	size_t name_to_id(const char *name);
+	int name_to_id(const char *name);
 
 public:
 	/* Detects file format, read header information, detects which data columns are available and at what index
